README.md
 Hi, I’m @peres69

A Norwegian Photographer, based in Sétubal, in Portugal.

I do my best learn portogeese. As an imigrant it's a call of duty, and to socialice: Language is the key.

I’m want to collaborate on art, photo/nft, Vlogg or listen to unusual stories peopple have to know. I hope to team up with some people to work with, too spread the word, the art and all kind of life. Present as vlogg, pics and text.